+++ app.go
@@ -4,9 +4,9 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/piheta/sept/infra/auth"
-	"github.com/piheta/sept/infra/db"
-	"github.com/piheta/sept/models"
+	"github.com/piheta/sept/backend/auth"
+	"github.com/piheta/sept/backend/db"
+	"github.com/piheta/sept/backend/models"
 )
 
 type App struct {
