+++ sept_data/README.md
@@ -1,0 +1,9 @@
+# Sept Data
+
+Here the user's salt file and Database are stored.
+These files are generated after the user registers and logs in.
+
+Multiple users can log into sept on one device. Each user will have their own database.
+The Database stores users, chats and messages that the user is interacting with. 
+
+Currently the database is not encrypted.
